bot/requirements.txt
pywikibot
requests
Pillow
Flask